Images with height == width would create full size thumbnails, fixed!
This was a fairly tough one to debug..
The left side of the expression in the while loop would evaluate to 0
which would cause it to skip the logic inside the while loop body.
Because of this, we would end up using the original height and width
as the resolution for the thumbnail.