CSS edits
All background color is now the same, meaning that the main node
and the container node can not be distinguished and site looks more clean.

Photo stream buttons are moved to the top so that users do not have to scroll all the
way to the bottom to "reach" them. Images in the stream are inside a scroll box.
We remove the footer section entirely.

Buttons are now grey and not blue.

Lightbox buttons are fainter and background is darker.